letmecreate: Add dependency to kmod-spidev
spidev is not enabled in the kernel and it is not part of
the default packages in OpenWRT. LetMeCreate relies on
the presence of /dev/spidev0.2 and /dev/spidev0.3 to use
the SPI bus. These files are created by spidev.
This commit ensures that selecting LetMeCreate will also
select kmod-spidev so that any use of SPI will work on the
board.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>